+++ app/components/Entity/components/AnalysisMethod/components/Workflow/types.ts
@@ -1,0 +1,7 @@
+import { Workflow } from "../../../../../../apis/catalog/brc-analytics-catalog/common/entities";
+
+export interface Props {
+  geneModelUrl: string | null;
+  genomeVersionAssemblyId: string;
+  workflow: Workflow;
+}