+++ app/components/Entity/components/AnalysisMethod/components/Workflow/workflow.styles.ts
@@ -1,0 +1,10 @@
+import styled from "@emotion/styled";
+import { Grid2 } from "@mui/material";
+
+export const StyledGrid2 = styled(Grid2)`
+  padding: 20px;
+
+  .MuiButton-root {
+    align-self: flex-start;
+  }
+`;